Fix various compiler warnings for moldft target
- Fixed `-Wcomment` in `src/madness/chem/xcfunctional.h`
- Fixed `-Wparentheses` in `src/madness/chem/PNOStructures.cpp`
- Fixed `-Wsign-compare` in `src/madness/mra/mra.h` and `src/madness/chem/CCStructures.h`
- Fixed `-Wunused-but-set-variable` in `src/madness/mra/leafop.h`
- Fixed `-Wmaybe-uninitialized` in `src/madness/world/cloud.h`

+++ src/madness/chem/CCStructures.h
@@ -264,7 +264,7 @@ struct Pairs {
 
     static Pairs vector2pairs(const std::vector<T>& argument, const PairVectorMap map) {
         Pairs<T> pairs;
-        for (int i=0; i<argument.size(); ++i) {
+        for (size_t i=0; i<argument.size(); ++i) {
             pairs.insert(map.map[i].first,map.map[i].second,argument[i]);
         }
         return pairs;
